split transformer and encoding into different modules
