+++ ai-service/src/ai_service/embeddings/encoding.py
@@ -1,57 +1,12 @@
 import logging
-from functools import lru_cache
 from typing import cast
 from sentence_transformers import SentenceTransformer
-from ai_service import errors, utils, constants
-import torch
+from ai_service import errors
+
+from .transformer import get_model
 
 
 logger = logging.getLogger(__name__)
-
-
-def _get_device() -> str:
-    """
-    Auto-detect the best available device for embedding computation.
-
-    Returns:
-        Device string: 'cuda', 'mps', or 'cpu'
-    """
-
-    if torch.cuda.is_available():
-        return "cuda"
-    elif hasattr(torch.backends, "mps") and torch.backends.mps.is_available():
-        return "mps"
-    else:
-        return "cpu"
-
-
-@lru_cache(maxsize=1)
-def _get_model() -> SentenceTransformer:
-    """
-    Singleton accessor; loads the embedding model on first call only.
-    Implements safe optimizations for code understanding tasks.
-
-    Returns:
-        Loaded SentenceTransformer model optimized for code embeddings.
-
-    Raises:
-        EmbeddingError: If model loading fails.
-    """
-    model_name = utils.get_env_var(constants.EMBEDDING_MODEL)
-    device = _get_device()
-
-    try:
-        logger.info(f"Loading embedding model: {model_name}")
-        logger.info(f"Device: {device}")
-
-        return SentenceTransformer(
-            model_name_or_path=model_name,
-            device=device,
-            trust_remote_code=False,  # Might be required for some code models
-            cache_folder=None,  # Use default cache location
-        )
-    except Exception as e:
-        raise errors.EmbeddingError.model_load_failed(model_name, e) from e
 
 
 def _encode_texts(
@@ -79,7 +34,7 @@ def _encode_texts(
     """
     if not texts or all(not text.strip() for text in texts):
         raise errors.EmbeddingError.empty_input()
-    model: SentenceTransformer = _get_model()
+    model: SentenceTransformer = get_model()
 
     try:
         # Use appropriate encoding method based on context
@@ -97,6 +52,17 @@ def _encode_texts(
         elif not is_query and hasattr(model, "encode_document"):
             logger.debug(f"Encoding {len(texts)} documents using encode_document")
             embeddings = model.encode_document(  # type: ignore
+                texts,
+                convert_to_numpy=True,
+                normalize_embeddings=True,
+                batch_size=batch_size,
+                precision="float32",
+                show_progress_bar=show_progress_bar,
+                device=None,  # Use model's configured device
+            )
+        else:
+            logger.debug(f"Encoding {len(texts)} texts using default method")
+            embeddings = model.encode(  # type: ignore
                 texts,
                 convert_to_numpy=True,
                 normalize_embeddings=True,
