+++ ai-service/src/ai_service/handlers/ingest.py
@@ -8,7 +8,7 @@ from ai_service import (
     errors,
     project_ingestor,
 )
-from ai_service.embeddings import embedder
+from ai_service.embeddings import encoding
 
 logger = logging.getLogger(__name__)
 router = APIRouter()
@@ -55,7 +55,7 @@ def ingest_github_project(canonical_github_url: str) -> None:
 
         if code_snippets:
             # Batch embed all documents at once for better performance
-            embeddings = embedder.embed_documents(code_snippets)
+            embeddings = encoding.embed_documents(code_snippets)
 
             db.add_chunks(code_snippets, embeddings)
             logger.info(f"Stored {len(code_snippets)} code snippets in ChromaDB.")
